Restructuring submodule
